refactor(web): add ControlledConnectorSelector for form-bound connectors
The server and webdav data source forms repeated the same Controller ->
ConnectorSelector block, so bind it once alongside the other Controlled*
fields.
